vocab_streamlit_app/requirements.txt
numpy==1.23.5
pandas==1.5.3
plotly==5.14.1
requests==2.30.0
slackclient==2.9.4
SQLAlchemy==1.4.47
streamlit==1.23.1
psycopg2-binary